+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.0.1
+## 0.1.0
 
 ### Enhancements
 - Existing Data Dictionaries can be converted to JSON format from a CSV file
@@ -6,4 +6,10 @@
 - Added tests for JSON validity of variables and domains
 - Added test to check presence/validity of variable type
 - Added test to check if a domain referenced from a variable exists
-
+- Tests can now be added as a group, or on a test-by-test basis
+  - Add `require 'spout/tests'` to your individual tests, or to your `test_helper.rb`
+  - Include the all or some of the tests provided by Spout
+    - `include Spout::Tests` for all tests
+    - `include Spout::Tests::JsonValidation` to verify valid JSON formats of domains and variables
+    - `include Spout::Tests::VariableTypeValidation` to verify valid variable type for variables
+    - `include Spout::Tests::DomainExistenceValidation` to verify existence of domains referenced by variables